# td-bs-management-user/src/main/java/com/trainingdev/td_bs_management_user/dto/business/UserComment.java
package com.trainingdev.td_bs_management_user.dto.business;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class UserComment {
    private String id;
    private String name;
    private String profilePhoto;
}