# td-bs-management-user/src/main/java/com/trainingdev/td_bs_management_user/dto/output/ErrorResponse.java
package com.trainingdev.td_bs_management_user.dto.output;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class ErrorResponse {
    private String code;
    private String message;
}